Vector Policy

Tabular policies including the following:

VectorPolicy holds a vector of actions, one for each state, ordered according to state_index.

ValuePolicy holds a matrix of values for state-action pairs and chooses the action with the highest value at the given state

POMDPPolicies.VectorPolicy — Type

VectorPolicy{S,A}

A generic MDP policy that consists of a vector of actions. The entry at stateindex(mdp, s) is the action that will be taken in state s.

Fields

mdp::MDP{S,A} the MDP problem

act::Vector{A} a vector of size |S| mapping state indices to actions

POMDPPolicies.VectorSolver — Type

VectorSolver{A}

Solver for VectorPolicy. Doesn't do any computation - just sets the action vector.

Fields

act::Vector{A} the action vector

POMDPPolicies.ValuePolicy — Type

 ValuePolicy{P<:Union{POMDP,MDP}, T<:AbstractMatrix{Float64}, A}

A generic MDP policy that consists of a value table. The entry at stateindex(mdp, s) is the action that will be taken in state s. It is expected that the order of the actions in the value table is consistent with the order of the actions in act. If act is not explicitly set in the construction, act is ordered according to actionindex.

Fields

mdp::P the MDP problem

value_table::T the value table as a |S|x|A| matrix

act::Vector{A} the possible actions