Alpha Vector Policy

Alpha Vector Policy

Represents a policy with a set of alpha vectors. Can be constructed with AlphaVectorPolicy(pomdp, alphas, action_map), where alphas is either a vector of vectors or an |S| x |A| matrix. The action_map argument is a vector of actions with length equal to the number of alpha vectors. If this argument is not provided, ordered_actions is used to generate a default action map. function.jl

POMDPPolicies.AlphaVectorPolicy — Type.

AlphaVectorPolicy{P<:POMDP, A}

Represents a policy with a set of alpha vector

Constructor:

`AlphaVectorPolicy(pomdp::POMDP, alphas)`

alphas can be a matrix or a vector of vectors

Fields

pomdp::P the POMDP problem

alphas::Vector{Vector{Float64}} the list of alpha vectors

action_map::Vector{A} a list of action corresponding to the alpha vectors

source