Stochastic Policies

Types for representing randomized policies:

StochasticPolicy samples actions from an arbitrary distribution.

UniformRandomPolicy samples actions uniformly (see RandomPolicy for a similar use)

CategoricalTabularPolicy samples actions from a categorical distribution with weights given by a ValuePolicy.

POMDPPolicies.StochasticPolicy — Type

StochasticPolicy{D, RNG <: AbstractRNG}

Represents a stochastic policy. Action are sampled from an arbitrary distribution.

Constructor:

`StochasticPolicy(distribution; rng=Random.GLOBAL_RNG)`

Fields

distribution::D

rng::RNG a random number generator

POMDPPolicies.CategoricalTabularPolicy — Type

CategoricalTabularPolicy

represents a stochastic policy sampling an action from a categorical distribution with weights given by a ValuePolicy

constructor:

CategoricalTabularPolicy(mdp::Union{POMDP,MDP}; rng=Random.GLOBAL_RNG)

Fields

stochastic::StochasticPolicy

value::ValuePolicy